+++ Changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 ## Alpha 0.1.1
-In development
+Released on 24th March 2018
 * Fixed a crash in compiled mods caused when no items or no blocks were registered
 * Removed the "climbable" attribute from block
 * Added new attributes to block: flammable, burn_forever, redstone_signal and gravity